More updates for Windows builds
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@825139 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/decaf/lang/Thread.cpp
@@ -621,7 +621,7 @@ void Thread::park( Thread* thread, long long mills, long long nanos ) {
         if( mills == 0 && nanos == 0 ) {
             thread->properties->mutex.wait();
         } else {
-            thread->properties->mutex.wait( mills, nanos );
+            thread->properties->mutex.wait( mills, (int)nanos );
         }
 
         thread->properties->parked = false;
